Add files via upload
I created a Python version of the classic Space Invaders arcade game using the Pygame library. In the game, the player controls a spaceship that moves left and right at the bottom of the screen and shoots lasers upward to destroy waves of alien invaders. I implemented collision detection so the lasers can hit and remove enemies, and added a scoring system to track progress. The game gets more challenging as you go, with faster enemy movements. I also included basic sound effects and animations to make the gameplay more engaging. It was a great way to practice using loops, classes, and event handling in Python.